Restructures files
